refactor: update names for issues fetcher
